+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.11.4
   * Allow unset field end check for empty value by php functions `unset` and `isset`
+  * Attach behaviors in mapping
 
 ## 1.11.3 (2014-12-31)
   * Fixed bug with default values of collection options